Fix bug when password is stored in cleartext [VPNLINUX-804]
A MR was initially made by a community user named @Anonymous941 on our proton-keyring-linux GitHub repo with the fix provided by another community user @Wizzerinus,
but it makes more sense for this change to be here as we don't know if any other backends
would possibly experience the same issue or not as of yet.

Problem description: The main problem here is when the keyring
has no password it does not seem to properly parse new line characters upon reading data from the keyring.
Though when the password for the keyring has been set, all new lines are properly parse and not treated as literal characters.
For this reason we need to ensure that we properly parse it.

Note: This is mostly a workaround/hack due to a bug in gnome-keyring, see https://discourse.gnome.org/t/possible-bug-or-feature-storing-getting-data-keyring-protected-vs-unprotected-keyring/20312